Problem: 'basic' module is still too big
Solution: extract numbers-related functionality into its own module
